feat(frontend): make LoaderWallets component reusable
